Keyboard layout conflict
US and UK keyboard layouts are different and as such the * amd # keys that were being sent were coming up as something else. Even after changing the keyboard region on the RaspberryPi, it would still not work.  So, i've changed the keys to known usable.  If you're using a Windows machine, you probably wont have this problem.

Readme Updated.